Restructure measurement.
